tools/tts/pyproject.toml
[project]
name = "tts"
version = "0.1.0"
description = "Local TTS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kokoro-onnx>=0.2.9",
    "nvidia-cudnn-cu12>=9.6.0.74",
    "onnxruntime-gpu>=1.20.1",
    "soundfile>=0.13.0",
]
